src/requirements-dev.txt
# Development dependencies for testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
coverage>=7.3.0
factory-boy>=3.3.0
requests>=2.31.0

